requirements.txt
# Core dependencies
numpy>=1.21.0
requests>=2.28.0
python-dotenv>=1.0.0

# Audio and transcription
torch>=2.0.0
nemo-toolkit[asr]>=1.20.0
sounddevice>=0.4.6
soundfile>=0.12.0
librosa>=0.10.0

# LangChain and OpenAI
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-core>=0.1.0
openai>=1.0.0

# Vector database and embeddings
pinecone>=2.0.0
sentence-transformers>=2.2.0
einops>=0.7.0  # Required for nomic embedding model
transformers>=4.30.0  # Required for embedding models
tokenizers>=0.13.0  # Required for fast tokenization
safetensors>=0.3.1  # For safe model loading
huggingface-hub>=0.16.0  # For downloading models

# Database
supabase>=2.0.0
psycopg>=3.1.0
psycopg2-binary>=2.9.0

# Data processing
lxml>=4.9.0
orjson>=3.9.0
tqdm>=4.65.0

# Security
bcrypt>=4.0.0

# Web UI
streamlit>=1.28.0
streamlit-webrtc>=0.47.0
av>=10.0.0
pyaudio>=0.2.11
